requirements.txt
Django==2.2
djangorestframework==3.12.2
requests==2.31.0
django-environ==0.11.2
pytest==8.1.1
pytest-django==4.8.0